undo: Move undo stack references under data.undo
Undo is spread across several top-level data modules and one large
checkpoint module. Its stack positions are stored as Git references, which
are named pointers to checkpoint commits, in data.undo_refs.

Keeping every undo module directly under data does not show which functions
are internal storage details and which functions commands may call.
Reference updates, file capture, restoration, and stack movement are
therefore hard to constrain independently.

This commit begins forming the undo package by moving the Git reference
names and update helpers under data.undo and updating session and
checkpoint callers.

Later commits will move worktree capture and restoration before snapshot
storage and conflict checks leave the large checkpoint module.
