+++ src/git_stage_batch/data/undo/__init__.py
@@ -1,0 +1,3 @@
+"""Undo and redo subsystem."""
+
+from __future__ import annotations
